ml-api/requirements.txt
flask
gunicorn
pandas
numpy<2.0.0
scikit-learn
xgboost
requests
tensorflow-cpu